requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
astroid==2.13.3
astunparse==1.6.3
black==22.12.0
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
faiss-cpu==1.7.3
Flask==2.2.2
Flask-RESTful==0.3.9
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
huggingface-hub==0.12.1
itsdangerous==2.1.2
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.1
oauthlib==3.2.2
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydocstyle==6.3.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.26.1
urllib3==1.26.14
Werkzeug==2.2.3
